PXT-179 Add a if_exists parameter to create/add APIs
This commit adds a new parameter - if_exists - for the
add_embedding_index API which can be used to handle scenarios
where a named index may already exist. It can direct the API
to not add the index and return, or raise it as an error,
or replace the existing index with a new one.

Also fixed a bug in drop_index where we were not removing
the index entry from the index metadata.

Testing: improved existing testcases and added new
testcases for the new parameters and the bug repro.