PXT-382: Bug: Similarity queries do not work correctly with ReloadTester
When there are multiple embedding indexes on a column, the similarity
query must specify the index to use. When such a query is executed
after reloading metadata via the ReloadTester APIs, it fails.

RCA: the index name is passed to the SimilarityExpr as a runtime parameter
and is not serialized to persistence as part of the SimilarityExpr. Therefore,
on reloading from persistence, the index name information is lost and is
set as None. Since there are multiple indexes, the SimilarityExpr fails
with an error.

This commit fixes SimilarityExpr to include the index used.

Testing: added coverage for the changes. test repro that failed before,
passes with the fix.